need to split the userAdd(Ln: 809 in imp) into a seperate verification function
-FilePointerReadWrite -Done
-get Reg validation -Done
-get Rac validation -Done
-User add verification -Done

-fix region/city for user add func
-Printf format -Pending
-Variable names -Pending

+++ information.h
@@ -46,7 +46,7 @@ void displayAll(struct node *head);
 void displayTown(struct node *head, const char *town);
 void displayRace(struct node *head, const char *race);
 
-void displayTestedAboveThresh(struct node *head, float lb);
+void displayTestedAboveThresh(struct node *head, int lb);
 
 void purgeData();
 void removeAtFront(struct node **head);
